Move RecognizerExecutor.cs and BuiltInRecognizer.cs into Text-Grab.Core
Continuing Phase 1. Scoped to RecognizerExecutor only this round -
PatternExecutor stays in the app since PatternItem.GetAll()/GetByName()
are wired to the settings singleton, which would cascade into a wider
refactor (splitting PatternItem, updating ~6 UI call sites).

Both RecognizerExecutor and PatternExecutor called into
GrabTemplateExecutor.ExtractMatchesByMode, an internal pure method buried
in an otherwise WPF/OCR-coupled 660-line class. Extracted just that
method (+ its ExtractByIndices helper) into a new Text-Grab.Core
MatchModeSelector so RecognizerExecutor doesn't need to reach back into
the app; GrabTemplateExecutor and PatternExecutor now call the Core
version too, removing the duplicate logic instead of leaving it behind.

Added Microsoft.Recognizers.Text.* package references (+ the existing
NuGet.CommandLine exclusion workaround) to Text-Grab.Core, and dropped
the now-redundant direct references from Text-Grab.csproj since they
flow in transitively through the project reference.
