+++ Text-Grab.Core/Models/BuiltInRecognizer.cs
