+++ Text-Grab/Text-Grab.csproj
@@ -104,11 +104,6 @@
         <PackageReference Include="Magick.NET.SystemDrawing" Version="8.0.25" />
         <PackageReference Include="Magick.NET.SystemWindowsMedia" Version="8.0.25" />
         <PackageReference Include="Markdig" Version="1.3.2" />
-        <PackageReference Include="Microsoft.Recognizers.Text.Choice" Version="1.8.13" />
-        <PackageReference Include="Microsoft.Recognizers.Text.DateTime" Version="1.8.13" />
-        <PackageReference Include="Microsoft.Recognizers.Text.Number" Version="1.8.13" />
-        <PackageReference Include="Microsoft.Recognizers.Text.NumberWithUnit" Version="1.8.13" />
-        <PackageReference Include="Microsoft.Recognizers.Text.Sequence" Version="1.8.13" />
         <PackageReference Include="Microsoft.Toolkit.Uwp.Notifications" Version="7.1.3" />
         <!-- Windows App SDK (stable) powers the existing on-device OCR / image-description / text AI features. -->
         <PackageReference Include="Microsoft.WindowsAppSDK.AI" Version="2.4.4" />
